+++ apps/decodex/src/orchestrator/lane_control.rs
@@ -1,33 +1,34 @@
 use std::{
 	collections::HashSet,
-	io::Error,
 	path::{Path, PathBuf},
-	process, thread,
-	time::{Duration, Instant},
+	time::Duration,
 };
 
-use libc::{ESRCH, SIGKILL, SIGTERM, c_int, pid_t};
 use serde::Serialize;
 use serde_json::Value;
+
+mod interrupt;
+mod render;
+mod steer;
+
+use self::{
+	interrupt::{
+		attempt_hard_lane_interrupt, attempt_soft_lane_interrupt, lane_interrupt_next_action,
+	},
+	render::{render_lane_inspect_report, render_lane_interrupt_report},
+	steer::{attempt_lane_steer, validate_lane_steer_request},
+};
 
 use crate::{
 	config::ServiceConfig,
 	orchestrator::{
-		self, ChildRunRef, DEFAULT_STATUS_RUN_LIMIT, LaneSteerReport, LaneSteerRequest,
-		OperatorRunStatus, OperatorStatusSnapshot,
+		self, DEFAULT_STATUS_RUN_LIMIT, LaneSteerReport, LaneSteerRequest, OperatorRunStatus,
+		OperatorStatusSnapshot,
 	},
 	prelude::{Result, eyre},
-	run_control::{
-		self, LaneControlInterruptRequest, LaneControlInterruptRequestInput,
-		LaneControlInterruptResponse, LaneControlResponseStatus, LaneControlSteerRequest,
-		LaneControlSteerRequestInput, LaneControlSteerResponse, LaneControlSteerResponseStatus,
-	},
+	run_control::{LaneControlInterruptResponse, LaneControlResponseStatus},
 	runtime,
-	state::{
-		RUN_CONTROL_ACTION_ACCEPTED, RUN_CONTROL_ACTION_COMPLETED, RUN_CONTROL_ACTION_FAILED,
-		RUN_CONTROL_ACTION_FALLBACK, RUN_CONTROL_ACTION_TIMED_OUT, RunControlActionReceipt,
-		RunControlActionRequest, StateStore,
-	},
+	state::{RunControlActionReceipt, StateStore},
 };
 
 pub(crate) const DEFAULT_STEER_RESULT_WAIT_TIMEOUT: Duration = Duration::from_secs(10);
@@ -415,10 +416,9 @@ fn soft_interrupt_allows_hard_fallback(
 	}
 
 	match soft.status.as_str() {
-		"pending" | "failed" | "unavailable" => {
+		"pending" | "failed" | "unavailable" =>
 			soft.error_class.as_deref() != Some("lane_not_active")
-				|| run.process_id.is_some() && run.process_alive != Some(false)
-		},
+				|| run.process_id.is_some() && run.process_alive != Some(false),
 		"rejected" => soft.error_class.as_deref() == Some("run_lease_missing"),
 		_ => false,
 	}
@@ -586,522 +586,6 @@ fn lane_control_operator_context(run: &OperatorRunStatus) -> Value {
 	})
 }
 
-fn attempt_lane_steer(
-	state_store: &StateStore,
-	project: &ServiceConfig,
-	run: &OperatorRunStatus,
-	request: &LaneSteerRequest<'_>,
-) -> Result<LaneSteerReport> {
-	let message_byte_count = request.message.len();
-	let message_line_count = lane_steer_message_line_count(request.message);
-	let context = lane_control_operator_context(run);
-	let metadata = serde_json::json!({
-		"expectedTurnId": request.expected_turn_id,
-		"messageByteCount": message_byte_count,
-		"messageLineCount": message_line_count,
-	});
-	let receipt = state_store.resolve_run_control_action(RunControlActionRequest {
-		project_id: project.service_id(),
-		issue_id: &run.issue_id,
-		run_id: &run.run_id,
-		attempt_number: run.attempt_number,
-		thread_id: run.thread_id.as_deref(),
-		turn_id: Some(request.expected_turn_id),
-		source: request.source,
-		action: "steer",
-		timeout_ms: Some(i64::try_from(request.wait_timeout.as_millis()).unwrap_or(i64::MAX)),
-		metadata: Some(&metadata),
-		context: Some(&context),
-	})?;
-
-	if receipt.outcome() != RUN_CONTROL_ACTION_ACCEPTED {
-		return Ok(lane_steer_report_from_rejected_receipt(
-			request.issue,
-			run,
-			&receipt,
-			request.expected_turn_id,
-			message_byte_count,
-			message_line_count,
-		));
-	}
-
-	let Some(worktree_path) = absolute_lane_worktree_path(project, state_store, run)? else {
-		eyre::bail!("Lane steer was accepted without a current lane worktree.");
-	};
-	let Some(thread_id) = run.thread_id.as_deref() else {
-		eyre::bail!("Lane steer was accepted before the active app-server thread id was known.");
-	};
-	let control_request = LaneControlSteerRequest::new(LaneControlSteerRequestInput {
-		audit_record_id: receipt.audit_record_id(),
-		project_id: project.service_id(),
-		issue_id: &run.issue_id,
-		run_id: &run.run_id,
-		attempt_number: run.attempt_number,
-		thread_id,
-		expected_turn_id: request.expected_turn_id,
-		source: request.source,
-		message: request.message,
-	});
-	let request_path = run_control::write_steer_request(&worktree_path, &control_request)?;
-
-	state_store.append_private_execution_event(
-		project.service_id(),
-		&run.issue_id,
-		&run.run_id,
-		run.attempt_number,
-		"lane_control/steer/requested",
-		serde_json::json!({
-			"requestId": control_request.request_id,
-			"source": request.source,
-			"method": "turn/steer",
-			"expectedTurnId": request.expected_turn_id,
-			"messageByteCount": control_request.message_byte_count,
-			"messageLineCount": control_request.message_line_count,
-		}),
-	)?;
-
-	match run_control::wait_for_steer_response(
-		&worktree_path,
-		&run.run_id,
-		&control_request.request_id,
-		request.wait_timeout,
-	)? {
-		Some(response) => {
-			let outcome = match &response.status {
-				LaneControlSteerResponseStatus::Delivered => RUN_CONTROL_ACTION_COMPLETED,
-				LaneControlSteerResponseStatus::Failed
-				| LaneControlSteerResponseStatus::Rejected => RUN_CONTROL_ACTION_FAILED,
-			};
-
-			state_store.record_run_control_action_outcome(
-				&receipt,
-				outcome,
-				&response.classification,
-			)?;
-
-			Ok(lane_steer_report_from_response(
-				request.issue,
-				run,
-				&receipt,
-				&control_request,
-				&request_path,
-				response,
-			))
-		},
-		None => {
-			state_store.record_run_control_action_outcome(
-				&receipt,
-				RUN_CONTROL_ACTION_TIMED_OUT,
-				"steer_response_pending",
-			)?;
-
-			Ok(lane_steer_report_pending(
-				request.issue,
-				run,
-				&receipt,
-				&control_request,
-				&request_path,
-			))
-		},
-	}
-}
-
-fn lane_steer_report_from_rejected_receipt(
-	issue: &str,
-	run: &OperatorRunStatus,
-	receipt: &RunControlActionReceipt,
-	expected_turn_id: &str,
-	message_byte_count: usize,
-	message_line_count: usize,
-) -> LaneSteerReport {
-	LaneSteerReport {
-		project_id: receipt.project_id().to_owned(),
-		issue_id: receipt.issue_id().to_owned(),
-		issue_identifier: run.issue_identifier.clone().or_else(|| Some(issue.to_owned())),
-		run_id: receipt.run_id().to_owned(),
-		attempt_number: receipt.attempt_number(),
-		thread_id: receipt.current_thread_id().map(str::to_owned),
-		expected_turn_id: expected_turn_id.to_owned(),
-		current_turn_id: receipt.current_turn_id().map(str::to_owned),
-		response_turn_id: None,
-		audit_record_id: receipt.audit_record_id(),
-		request_id: String::new(),
-		request_path: None,
-		outcome: receipt.outcome().to_owned(),
-		reason: receipt.reason().to_owned(),
-		failure_class: lane_steer_failure_class_for_reason(receipt.reason()).map(str::to_owned),
-		delivery_status: String::from("rejected"),
-		message_byte_count,
-		message_line_count,
-	}
-}
-
-fn lane_steer_report_from_response(
-	issue: &str,
-	run: &OperatorRunStatus,
-	receipt: &RunControlActionReceipt,
-	request: &LaneControlSteerRequest,
-	request_path: &Path,
-	response: LaneControlSteerResponse,
-) -> LaneSteerReport {
-	let outcome = match &response.status {
-		LaneControlSteerResponseStatus::Delivered => RUN_CONTROL_ACTION_COMPLETED,
-		LaneControlSteerResponseStatus::Failed | LaneControlSteerResponseStatus::Rejected => {
-			RUN_CONTROL_ACTION_FAILED
-		},
-	};
-
-	LaneSteerReport {
-		project_id: receipt.project_id().to_owned(),
-		issue_id: receipt.issue_id().to_owned(),
-		issue_identifier: run.issue_identifier.clone().or_else(|| Some(issue.to_owned())),
-		run_id: receipt.run_id().to_owned(),
-		attempt_number: receipt.attempt_number(),
-		thread_id: Some(response.thread_id.clone()),
-		expected_turn_id: response.expected_turn_id.clone(),
-		current_turn_id: response.current_turn_id.clone(),
-		response_turn_id: response.response_turn_id.clone(),
-		audit_record_id: receipt.audit_record_id(),
-		request_id: response.request_id.clone(),
-		request_path: Some(request_path.display().to_string()),
-		outcome: outcome.to_owned(),
-		reason: response.classification.clone(),
-		failure_class: response.error_class.clone(),
-		delivery_status: String::from("resolved"),
-		message_byte_count: request.message_byte_count,
-		message_line_count: request.message_line_count,
-	}
-}
-
-fn lane_steer_report_pending(
-	issue: &str,
-	run: &OperatorRunStatus,
-	receipt: &RunControlActionReceipt,
-	request: &LaneControlSteerRequest,
-	request_path: &Path,
-) -> LaneSteerReport {
-	LaneSteerReport {
-		project_id: receipt.project_id().to_owned(),
-		issue_id: receipt.issue_id().to_owned(),
-		issue_identifier: run.issue_identifier.clone().or_else(|| Some(issue.to_owned())),
-		run_id: receipt.run_id().to_owned(),
-		attempt_number: receipt.attempt_number(),
-		thread_id: Some(request.thread_id.clone()),
-		expected_turn_id: request.expected_turn_id.clone(),
-		current_turn_id: run.turn_id.clone(),
-		response_turn_id: None,
-		audit_record_id: receipt.audit_record_id(),
-		request_id: request.request_id.clone(),
-		request_path: Some(request_path.display().to_string()),
-		outcome: RUN_CONTROL_ACTION_ACCEPTED.to_owned(),
-		reason: String::from("queued_wait_timeout"),
-		failure_class: None,
-		delivery_status: String::from("queued"),
-		message_byte_count: request.message_byte_count,
-		message_line_count: request.message_line_count,
-	}
-}
-
-fn validate_lane_steer_request(request: &LaneSteerRequest<'_>) -> Result<()> {
-	if request.issue.trim().is_empty() {
-		eyre::bail!("Lane steer issue must not be empty.");
-	}
-	if request.run_id.trim().is_empty() {
-		eyre::bail!("Lane steer run id must not be empty.");
-	}
-	if request.expected_turn_id.trim().is_empty() {
-		eyre::bail!("Lane steer expected turn id must not be empty.");
-	}
-	if request.message.trim().is_empty() {
-		eyre::bail!("Lane steer message must not be empty.");
-	}
-	if request.source.trim().is_empty() {
-		eyre::bail!("Lane steer source must not be empty.");
-	}
-
-	Ok(())
-}
-
-fn lane_steer_failure_class_for_reason(reason: &str) -> Option<&'static str> {
-	match reason {
-		"turn_mismatch" | "stale_expected_turn_id" => Some("stale_expected_turn_id"),
-		"active_turn_not_steerable" => Some("active_turn_not_steerable"),
-		"app_server_turn_steer_timed_out" => Some("app_server_turn_steer_timed_out"),
-		"app_server_turn_steer_unsupported" => Some("app_server_turn_steer_unsupported"),
-		"app_server_turn_steer_failed" => Some("app_server_turn_steer_failed"),
-		"run_lease_control_channel_resolved" | "queued_wait_timeout" => None,
-		_ => Some("run_control_action_failed"),
-	}
-}
-
-fn lane_steer_message_line_count(message: &str) -> usize {
-	message.lines().count().max(usize::from(!message.is_empty()))
-}
-
-fn attempt_soft_lane_interrupt(
-	state_store: &StateStore,
-	project: &ServiceConfig,
-	run: &OperatorRunStatus,
-	force: bool,
-	reason: Option<&str>,
-	source: &str,
-) -> Result<LaneSoftInterruptReport> {
-	let Some(worktree_path) = absolute_lane_worktree_path(project, state_store, run)? else {
-		return Ok(LaneSoftInterruptReport::unavailable(
-			"worktree_missing",
-			"Soft interrupt requires the current lane worktree and run-control directory.",
-		));
-	};
-	let Some(thread_id) = run.thread_id.as_deref() else {
-		return Ok(LaneSoftInterruptReport::unavailable(
-			"thread_id_missing",
-			"Soft interrupt requires a recorded app-server thread id.",
-		));
-	};
-	let Some(turn_id) = run.turn_id.as_deref() else {
-		return Ok(LaneSoftInterruptReport::unavailable(
-			"turn_id_missing",
-			"Soft interrupt requires a recorded active app-server turn id.",
-		));
-	};
-
-	if !orchestrator::operator_run_counts_as_current_lane(run) {
-		return Ok(LaneSoftInterruptReport::unavailable(
-			"lane_not_active",
-			"Soft interrupt only targets current or live local lane runs.",
-		));
-	}
-
-	let receipt = resolve_soft_interrupt_control_action(
-		state_store,
-		project,
-		run,
-		thread_id,
-		turn_id,
-		source,
-	)?;
-
-	if receipt.outcome() != "accepted" {
-		return Ok(LaneSoftInterruptReport::from_control_rejection(&receipt));
-	}
-
-	let request = LaneControlInterruptRequest::new(LaneControlInterruptRequestInput {
-		project_id: project.service_id(),
-		issue_id: &run.issue_id,
-		run_id: &run.run_id,
-		attempt_number: run.attempt_number,
-		thread_id,
-		turn_id,
-		source,
-		reason,
-	});
-
-	run_control::write_interrupt_request(&worktree_path, &request)?;
-
-	state_store.append_private_execution_event(
-		project.service_id(),
-		&run.issue_id,
-		&run.run_id,
-		run.attempt_number,
-		"lane_control/interrupt/requested",
-		serde_json::json!({
-			"requestId": request.request_id,
-			"source": source,
-			"method": "turn/interrupt",
-			"force": force,
-			"reason": reason,
-		}),
-	)?;
-
-	match run_control::wait_for_interrupt_response(
-		&worktree_path,
-		&run.run_id,
-		&request.request_id,
-		LANE_INTERRUPT_RESPONSE_WAIT,
-	)? {
-		Some(response) => {
-			let outcome = match &response.status {
-				LaneControlResponseStatus::SoftDelivered => RUN_CONTROL_ACTION_COMPLETED,
-				LaneControlResponseStatus::SoftFailed => RUN_CONTROL_ACTION_FAILED,
-				LaneControlResponseStatus::Rejected => RUN_CONTROL_ACTION_FAILED,
-			};
-
-			state_store.record_run_control_action_outcome(
-				&receipt,
-				outcome,
-				&response.classification,
-			)?;
-
-			Ok(LaneSoftInterruptReport::from_response(response))
-		},
-		None => {
-			state_store.record_run_control_action_outcome(
-				&receipt,
-				RUN_CONTROL_ACTION_TIMED_OUT,
-				"soft_interrupt_response_pending",
-			)?;
-
-			Ok(LaneSoftInterruptReport {
-				attempted: true,
-				available: true,
-				status: String::from("pending"),
-				classification: String::from("soft_interrupt_pending"),
-				method: String::from("turn/interrupt"),
-				request_id: Some(request.request_id),
-				message: String::from(
-					"Soft interrupt request was written, but the app-server child has not recorded a response yet.",
-				),
-				error_class: Some(String::from("soft_interrupt_response_pending")),
-				protocol_summary: None,
-				response: None,
-			})
-		},
-	}
-}
-
-fn resolve_soft_interrupt_control_action(
-	state_store: &StateStore,
-	project: &ServiceConfig,
-	run: &OperatorRunStatus,
-	thread_id: &str,
-	turn_id: &str,
-	source: &str,
-) -> Result<RunControlActionReceipt> {
-	let context = lane_control_operator_context(run);
-
-	state_store.resolve_run_control_action(RunControlActionRequest {
-		project_id: project.service_id(),
-		issue_id: &run.issue_id,
-		run_id: &run.run_id,
-		attempt_number: run.attempt_number,
-		thread_id: Some(thread_id),
-		turn_id: Some(turn_id),
-		source,
-		action: "interrupt",
-		timeout_ms: Some(
-			i64::try_from(LANE_INTERRUPT_RESPONSE_WAIT.as_millis()).unwrap_or(i64::MAX),
-		),
-		metadata: None,
-		context: Some(&context),
-	})
-}
-
-fn attempt_hard_lane_interrupt(
-	state_store: &StateStore,
-	run: &OperatorRunStatus,
-	reason: Option<&str>,
-) -> Result<LaneHardInterruptReport> {
-	let Some(process_id) = run.process_id else {
-		return Ok(LaneHardInterruptReport::unavailable(
-			"process_id_missing",
-			"Hard interrupt fallback requires a recorded child process id.",
-		));
-	};
-
-	if process_id == process::id() {
-		eyre::bail!("Refusing to hard-interrupt the current Decodex process.");
-	}
-
-	let mut signals = Vec::new();
-	let mut sent_any_signal = false;
-
-	if send_lane_signal(process_id, SIGTERM)? {
-		sent_any_signal = true;
-
-		signals.push(String::from("SIGTERM"));
-	}
-	if sent_any_signal
-		&& !wait_for_lane_process_exit(process_id, LANE_HARD_INTERRUPT_TERM_WAIT)
-		&& send_lane_signal(process_id, SIGKILL)?
-	{
-		signals.push(String::from("SIGKILL"));
-	}
-
-	let process_alive_after = Some(orchestrator::process_is_alive(process_id));
-	let status = if process_alive_after == Some(false) {
-		"sent"
-	} else if sent_any_signal {
-		"still_alive"
-	} else {
-		"process_not_found"
-	};
-	let message = if sent_any_signal {
-		String::from("Hard interrupt fallback signaled the recorded child process.")
-	} else {
-		String::from("Hard interrupt fallback found no signalable child process.")
-	};
-
-	state_store.append_private_execution_event(
-		&run.project_id,
-		&run.issue_id,
-		&run.run_id,
-		run.attempt_number,
-		"lane_control/interrupt",
-		serde_json::json!({
-			"classification": "hard_interrupt_fallback",
-			"status": status,
-			"signals": signals,
-			"processId": process_id,
-			"processAliveAfter": process_alive_after,
-			"reason": reason,
-		}),
-	)?;
-
-	record_hard_interrupt_control_fallback(state_store, run, reason)?;
-
-	if sent_any_signal {
-		orchestrator::clear_orphaned_daemon_child_state(
-			state_store,
-			ChildRunRef {
-				issue_id: &run.issue_id,
-				run_id: &run.run_id,
-				attempt_number: run.attempt_number,
-			},
-			true,
-		)?;
-	}
-
-	Ok(LaneHardInterruptReport {
-		attempted: true,
-		status: String::from(status),
-		classification: String::from("hard_interrupt_fallback"),
-		signals,
-		process_id: Some(process_id),
-		process_alive_after,
-		message,
-		error_class: None,
-	})
-}
-
-fn record_hard_interrupt_control_fallback(
-	state_store: &StateStore,
-	run: &OperatorRunStatus,
-	_reason: Option<&str>,
-) -> Result<()> {
-	let context = lane_control_operator_context(run);
-	let receipt = state_store.resolve_run_control_action(RunControlActionRequest {
-		project_id: &run.project_id,
-		issue_id: &run.issue_id,
-		run_id: &run.run_id,
-		attempt_number: run.attempt_number,
-		thread_id: run.thread_id.as_deref(),
-		turn_id: run.turn_id.as_deref(),
-		source: "hard_interrupt_fallback",
-		action: "interrupt",
-		timeout_ms: None,
-		metadata: None,
-		context: Some(&context),
-	})?;
-
-	state_store.record_run_control_action_outcome(
-		&receipt,
-		RUN_CONTROL_ACTION_FALLBACK,
-		"hard_interrupt_fallback",
-	)?;
-
-	Ok(())
-}
-
 fn absolute_lane_worktree_path(
 	project: &ServiceConfig,
 	state_store: &StateStore,
@@ -1121,150 +605,4 @@ fn absolute_lane_worktree_path(
 	} else {
 		project.repo_root().join(worktree_path)
 	}))
-}
-
-fn send_lane_signal(process_id: u32, signal: c_int) -> Result<bool> {
-	let process_id = pid_t::try_from(process_id).map_err(|_error| {
-		eyre::eyre!("Recorded child process id is too large for this platform.")
-	})?;
-
-	if process_id <= 0 {
-		eyre::bail!("Recorded child process id must be positive.");
-	}
-
-	match unsafe { libc::kill(process_id, signal) } {
-		0 => Ok(true),
-		-1 if Error::last_os_error().raw_os_error() == Some(ESRCH) => Ok(false),
-		-1 => Err(Error::last_os_error().into()),
-		_ => Ok(false),
-	}
-}
-
-fn wait_for_lane_process_exit(process_id: u32, timeout: Duration) -> bool {
-	let started_at = Instant::now();
-
-	while started_at.elapsed() < timeout {
-		if !orchestrator::process_is_alive(process_id) {
-			return true;
-		}
-
-		thread::sleep(Duration::from_millis(100));
-	}
-
-	!orchestrator::process_is_alive(process_id)
-}
-
-fn lane_interrupt_next_action(
-	soft: &LaneSoftInterruptReport,
-	hard: Option<&LaneHardInterruptReport>,
-	force: bool,
-) -> String {
-	if let Some(hard) = hard {
-		return if hard.status == "unavailable" {
-			String::from("Hard fallback was unavailable; inspect the lane before retrying.")
-		} else if hard.status == "sent" || hard.status == "process_not_found" {
-			String::from(
-				"Inspect the lane to confirm the lease and dirty-worktree reconciliation state.",
-			)
-		} else {
-			String::from(
-				"The fallback signal did not stop the recorded process; inspect the host process before retrying.",
-			)
-		};
-	}
-
-	match soft.status.as_str() {
-		"delivered" => {
-			String::from("Inspect the lane until the app-server turn records completion.")
-		},
-		"pending" => {
-			if force {
-				String::from("Soft interrupt is pending; forced fallback was not attempted.")
-			} else {
-				String::from(
-					"Re-run inspect shortly, or retry interrupt with --force if operator intent is to kill the process.",
-				)
-			}
-		},
-		"rejected" => String::from(
-			"Inspect the lane identity before retrying; resolver rejection is not converted into hard fallback.",
-		),
-		"failed" | "unavailable" => String::from(
-			"Retry with --force only if operator intent is to use hard process-kill fallback.",
-		),
-		_ => String::from("Inspect the lane for the latest run status."),
-	}
-}
-
-fn render_lane_inspect_report(report: &LaneInspectReport) -> String {
-	let mut output = format!(
-		"Lane inspect for {} in project {} ({} run{})\n",
-		report.issue,
-		report.project_id,
-		report.matched_run_count,
-		if report.matched_run_count == 1 { "" } else { "s" }
-	);
-
-	for run in &report.runs {
-		output.push_str(&format!(
-			"- {} attempt {}: status={}, phase={}, runLease={}, owner={}, liveness={}\n",
-			run.run_id,
-			run.attempt_number,
-			run.status,
-			run.phase,
-			run.run_lease,
-			run.ownership_state,
-			run.execution_liveness
-		));
-		output.push_str(&format!(
-			"  laneControl: livenessState={}, policyState={}, terminalization={}, nextAction={}, conditions={}\n",
-			run.liveness_state,
-			run.policy_state,
-			run.terminalization_state,
-			run.lane_control_next_action,
-			if run.lane_control_conditions.is_empty() {
-				String::from("none")
-			} else {
-				run.lane_control_conditions.join(",")
-			}
-		));
-		output.push_str(&format!(
-			"  appServer: thread={}, turn={}, softInterruptAvailable={}\n",
-			run.thread_id.as_deref().unwrap_or("none"),
-			run.turn_id.as_deref().unwrap_or("none"),
-			run.soft_interrupt_available
-		));
-		output.push_str(&format!(
-			"  process: pid={}, alive={}, hardInterruptAvailable={} (requires --force)\n",
-			run.process_id.map_or_else(|| String::from("none"), |id| id.to_string()),
-			run.process_alive.map_or_else(|| String::from("unknown"), |alive| alive.to_string()),
-			run.hard_interrupt_available
-		));
-	}
-
-	output
-}
-
-fn render_lane_interrupt_report(report: &LaneInterruptReport) -> String {
-	let mut output = format!(
-		"Lane interrupt {} for run {}: {}\n",
-		report.classification, report.run_id, report.soft_interrupt.message
-	);
-
-	if let Some(hard) = &report.hard_interrupt {
-		output.push_str(&format!(
-			"Hard fallback {}: {} ({})\n",
-			hard.status,
-			hard.message,
-			if hard.signals.is_empty() {
-				String::from("no signals")
-			} else {
-				hard.signals.join(",")
-			}
-		));
-	}
-
-	output.push_str(&format!("Next action: {}\n", report.next_action));
-
-	output
 }
